Revert " [AMDGPU][WaveTransform] merge amd-staging into amd-feature/wave-transform (#1837)"
This reverts commit 4d25bcdfc39cd3901402e608525d766acb0bb593.

This merge has some issues, so reverting it.

+++ libc/test/src/math/exhaustive/exhaustive_test.h
@@ -169,8 +169,7 @@ struct LlvmLibcExhaustiveMathTest
               range_end = stop;
             }
             current_value = range_end;
-            int pc =
-                static_cast<int>(100.0 * (range_end - start) / (stop - start));
+            int pc = 100.0 * (range_end - start) / (stop - start);
             if (current_percent != pc) {
               new_percent = pc;
               current_percent = pc;
